+++ packages/harpstrata/src/harpstrata/HarpStrata/getHarpStrata/index.ts
@@ -1,4 +1,5 @@
 import type { HarpStrataProps, HarpStrata } from '../types'
+import { getRootPitch } from '../getRootPitch'
 import { getPozition } from '../../Pozition'
 import { getPitchMatrix, getPitch } from '../../Pitch'
 import { getIsActiveComplex } from '../../IsActive'
@@ -6,11 +7,11 @@ import { getDegreeMatrix } from '../../Degree'
 import { getApparatus } from '../../Apparatus'
 
 export const getHarpStrata = (props: HarpStrataProps): HarpStrata => {
-  const { apparatusId, pozitionId, keyPitchId, activeIds } = props
+  const { apparatusId, pozitionId, harpKeyId, activeIds } = props
 
   const apparatus = getApparatus(apparatusId)
   const pozition = getPozition(pozitionId)
-  const pitch = getPitch(keyPitchId)
+  const pitch = getPitch(harpKeyId)
 
   const { halfstepIndexMatrix } = apparatus
   const { root: pozitionRoot } = pozition
@@ -20,10 +21,15 @@ export const getHarpStrata = (props: HarpStrataProps): HarpStrata => {
   const pitchMatrix = getPitchMatrix(halfstepIndexMatrix, pitchId)
   const isActiveComplex = getIsActiveComplex({degreeMatrix, pitchMatrix, activeIds: activeIds })
 
+  const rootPitchId = getRootPitch(degreeMatrix, pitchMatrix)
+
   return {
     apparatus,
     degreeMatrix,
     pitchMatrix,
     isActiveComplex,
+    pozitionId,
+    rootPitchId,
+    harpKeyId,
   }
 }