+++ packages/harpstrata/src/harpstrata/HarpStrata/types/index.ts
@@ -7,7 +7,7 @@ import type { Apparatus, ApparatusIds } from '../../Apparatus'
 export type HarpStrataProps = {
   readonly apparatusId: ApparatusIds;
   readonly pozitionId: PozitionIds;
-  readonly keyPitchId: PitchIds;
+  readonly harpKeyId: PitchIds;
   readonly activeIds: ActiveIds;
 }
 
@@ -16,4 +16,7 @@ export type HarpStrata = {
   readonly degreeMatrix: DegreeMatrix;
   readonly pitchMatrix: PitchMatrix;
   readonly isActiveComplex: IsActiveComplex;
+  readonly pozitionId: PozitionIds;
+  readonly rootPitchId: PitchIds | undefined;
+  readonly harpKeyId: PitchIds;
 }