+++ packages/harpstrata/src/harpstrata/Pitch/index.ts
@@ -38,3 +38,5 @@ export const getPitchMatrix = (halfstepIndexMatrix: HalfstepIndexMatrix, keyPitc
 export type { Pitch, PitchRow, PitchMatrix } from './types'
 export { PitchIds } from './types'
 export { A, Bb, B, C, Db, D, Eb, E, F, Gb, G, Ab } from './constants'
+
+export { EXAMPLE_PITCH_MATRICES } from './testResources'